moved the asm ClassWriter subclass to analysis directory.
